Improve stability of fork
Instead of the forked service receiving the new socket fd over the
previous socket after forking, the new socket fd is received before
forking the process. This eliminates the need for the old service to
wait until the socket fd exchange has finished.

Funcionality to wait for the process to finished has been changed to
utilize the process ID (PID) instead of the Popen object. This allows
the forked process to relay it's new PID to the Client instance, after
which the Client instance can wait for this PID to close when closing
down.