Fix for proper process hierarchy
When the child process is the one to receive a new socket fd, the
process hierarchy of a single process continually forking does not show
a single process with many children, but instead a stack of children.
This fix ensures that the process hierarchy is consistent with what is
to be expected.